refactor: move sidebar sections to config with docs
